+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ### [0.9.1](https://github.com/openfga/dotnet-sdk/compare/v0.9.0...v0.9.1) (2026-01-26)
 
 - feat: add support for streamed list objects (#156)
-- feat: add support for an interface on the client  (#168)
+- feat: add support for an interface on the client (#168)
 
 ## v0.9.0